refactor(filter-multi-select): rename scss files to .module.scss
