+++ packages/select/src/FilterMultiSelect/components/ListBox/ListBox.module.scss
