+++ packages/select/src/FilterMultiSelect/components/ListBoxSection/ListBoxSection.module.scss
