+++ packages/select/src/FilterMultiSelect/components/MenuLayout/MenuFooter/MenuFooter.module.scss
