+++ packages/select/src/FilterMultiSelect/components/MenuLayout/MenuLoadingSkeleton/MenuLoadingSkeleton.module.scss
