+++ packages/select/src/FilterMultiSelect/components/MenuPopup/MenuPopup.tsx
@@ -3,7 +3,7 @@ import { useOverlay, DismissButton } from "@react-aria/overlays"
 import { FocusScope } from "@react-aria/focus"
 import { mergeProps } from "@react-aria/utils"
 import { useMenuTriggerContext } from "../../provider"
-import styles from "./MenuPopup.scss"
+import styles from "./MenuPopup.module.scss"
 
 export interface MenuPopupProps {
   isLoading?: boolean
