+++ packages/select/src/FilterMultiSelect/components/MultiSelectOption/MultiSelectOption.module.scss
