+++ packages/select/src/FilterMultiSelect/components/Trigger/FilterTriggerButton/FilterTriggerButton.module.scss
