+++ packages/select/src/FilterMultiSelect/components/Trigger/RemovableFilterTrigger/RemovableFilterTrigger.module.scss
