+++ packages/select/src/FilterMultiSelect/components/Trigger/TriggerButtonBase/TriggerButtonBase.module.scss
